requirements.txt
frontmatter==3.0.8
python-frontmatter==1.0.0
PyYAML==5.1
setuptools==75.6.0
